revert: fetch recovery state via listener
