+++ src/components/recovery/RecoveryContext/useRecoveryState.ts
@@ -1,6 +1,5 @@
-import { useCallback, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import type { Delay } from '@gnosis.pm/zodiac'
-import type { TransactionListItem } from '@safe-global/safe-gateway-typescript-sdk'
 
 import useAsync from '@/hooks/useAsync'
 import { useCurrentChain } from '@/hooks/useChains'
@@ -8,42 +7,15 @@ import useSafeInfo from '@/hooks/useSafeInfo'
 import { useWeb3ReadOnly } from '@/hooks/wallets/web3'
 import useIntervalCounter from '@/hooks/useIntervalCounter'
 import { getRecoveryState } from '@/services/recovery/recovery-state'
-import useTxHistory from '@/hooks/useTxHistory'
+import { useAppDispatch } from '@/store'
 import { isCustomTxInfo, isMultiSendTxInfo, isTransactionListItem } from '@/utils/transaction-guards'
 import { sameAddress } from '@/utils/addresses'
+import { addListener } from '@reduxjs/toolkit'
+import { txHistorySlice } from '@/store/txHistorySlice'
 import type { AsyncResult } from '@/hooks/useAsync'
 import type { RecoveryState } from '@/services/recovery/recovery-state'
 
 const REFRESH_DELAY = 5 * 60 * 1_000 // 5 minutes
-
-export function _shouldFetchRecoveryState(
-  delayModifiers?: Array<Delay>,
-  txHistory?: Array<TransactionListItem>,
-): delayModifiers is Array<Delay> {
-  if (!delayModifiers || delayModifiers.length === 0) {
-    return false
-  }
-
-  // Transactions cleared on new Safe session
-  if (!txHistory) {
-    return true
-  }
-
-  const [latestTx] = txHistory.filter(isTransactionListItem)
-
-  if (!latestTx) {
-    return false
-  }
-
-  const { txInfo } = latestTx.transaction
-
-  const isDelayModiferTx = delayModifiers.some((delayModifier) => {
-    return isCustomTxInfo(txInfo) && sameAddress(txInfo.to.value, delayModifier.address)
-  })
-
-  // Multiple Delay Modifier settings changes are batched into a MultiSend
-  return isDelayModiferTx || isMultiSendTxInfo(txInfo)
-}
 
 export function useRecoveryState(delayModifiers?: Array<Delay>): {
   data: AsyncResult<RecoveryState>
@@ -52,7 +24,7 @@ export function useRecoveryState(delayModifiers?: Array<Delay>): {
   const web3ReadOnly = useWeb3ReadOnly()
   const chain = useCurrentChain()
   const { safe, safeAddress } = useSafeInfo()
-  const txHistory = useTxHistory()
+  const dispatch = useAppDispatch()
 
   // Reload recovery data every REFRESH_DELAY
   const [counter] = useIntervalCounter(REFRESH_DELAY)
@@ -65,20 +37,18 @@ export function useRecoveryState(delayModifiers?: Array<Delay>): {
 
   const data = useAsync<RecoveryState>(
     () => {
-      if (
-        _shouldFetchRecoveryState(delayModifiers, txHistory.page?.results) &&
-        chain?.transactionService &&
-        web3ReadOnly
-      ) {
-        return getRecoveryState({
-          delayModifiers,
-          transactionService: chain.transactionService,
-          safeAddress,
-          provider: web3ReadOnly,
-          chainId: safe.chainId,
-          version: safe.version,
-        })
+      if (!delayModifiers || delayModifiers.length === 0 || !chain?.transactionService || !web3ReadOnly) {
+        return
       }
+
+      return getRecoveryState({
+        delayModifiers,
+        transactionService: chain.transactionService,
+        safeAddress,
+        provider: web3ReadOnly,
+        chainId: safe.chainId,
+        version: safe.version,
+      })
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [
@@ -90,10 +60,53 @@ export function useRecoveryState(delayModifiers?: Array<Delay>): {
       safeAddress,
       safe.chainId,
       safe.version,
-      txHistory.page?.results,
     ],
     false,
   )
+
+  // Reload recovery data when a Delay Modifier is interacted with
+  useEffect(() => {
+    if (!delayModifiers || delayModifiers.length === 0) {
+      return
+    }
+
+    // We leverage a listener instead of useAsync dependencies because there are
+    // that need be loaded before we can initially fetch the recovery state
+    const listener = dispatch(
+      addListener({
+        // Listen to history polls (only occuring when the txHistoryTag changes)
+        actionCreator: txHistorySlice.actions.set,
+        effect: (action) => {
+          // Get the most recent transaction
+          const [latestTx] = action.payload.data?.results.filter(isTransactionListItem) ?? []
+
+          if (!latestTx) {
+            return
+          }
+
+          const { txInfo } = latestTx.transaction
+
+          const isDelayModiferTx = delayModifiers.some((delayModifier) => {
+            return isCustomTxInfo(txInfo) && sameAddress(txInfo.to.value, delayModifier.address)
+          })
+
+          // Refetch if the most recent transaction was with a Delay Modifier or MultiSend
+          // (Multiple Delay Modifier settings changes are batched into a MultiSend)
+          if (isDelayModiferTx || isMultiSendTxInfo(txInfo)) {
+            refetch()
+          }
+        },
+      }),
+    )
+
+    // Types are incorrect, but this ensures type safety
+    const unsubscribe =
+      listener instanceof Function
+        ? (listener as unknown as typeof listener.payload.unsubscribe)
+        : listener.payload.unsubscribe
+
+    return unsubscribe
+  }, [safe.chainId, delayModifiers, refetch, dispatch])
 
   return { data, refetch }
 }
